requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
bcrypt==4.3.0
black==25.1.0
cffi==1.17.1
click==8.1.8
contextlib2==21.6.0
cryptography==44.0.2
Django==5.1.5
django-cors-headers==4.7.0
django-ninja==1.3.0
django-ninja-extra==0.22.5
django-ninja-jwt==5.3.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
injector==0.22.0
mypy-extensions==1.0.0
mysqlclient==2.2.7
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.9.0
setuptools==75.1.0
sqlparse==0.5.3
typing_extensions==4.12.2
wheel==0.44.0
